#91 Restructured frontend folder structure
