+++ frontend/src/components/Login/Login.scss
@@ -1,4 +1,4 @@
-@import "../styles/prelude.scss";
+@import "../../styles/prelude.scss";
 
 .login-field{
     width: 100%;
