+++ frontend/src/components/Post/Post.scss
@@ -1,4 +1,4 @@
-@import "../styles/prelude.scss";
+@import "../../styles/prelude.scss";
 
 .post {
   display: flex;
