+++ frontend/src/components/Register/Register.scss
@@ -1,4 +1,4 @@
-@import "../styles/prelude.scss";
+@import "../../styles/prelude.scss";
 
 input{
     margin-bottom: 5px;
