+++ frontend/src/components/User/User.scss
@@ -1,4 +1,4 @@
-@import "../styles/prelude.scss";
+@import "../../styles/prelude.scss";
 
 .follow-button{
     color: rgb(25, 136, 240);
